NU5127 should work with buildTransitive
