+++ src/NuGet.Core/NuGet.Packaging/GlobalSuppressions.cs
@@ -88,7 +88,6 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Build", "CA1822:Member FindMatchingPackageNode does not access instance data and can be marked as static (Shared in VisualBasic)", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.PackagesConfigWriter.FindMatchingPackageNode(NuGet.Packaging.PackageReference,System.Xml.Linq.XElement)~System.Xml.Linq.XElement")]
 [assembly: SuppressMessage("Build", "CA1822:Member SortPackageNodes does not access instance data and can be marked as static (Shared in VisualBasic)", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.PackagesConfigWriter.SortPackageNodes(System.Xml.Linq.XElement)")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'IEnumerable<PackagingLogMessage> IconUrlDeprecationWarning.Validate(PackageArchiveReader builder)', validate parameter 'builder' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.Rules.IconUrlDeprecationWarning.Validate(NuGet.Packaging.PackageArchiveReader)~System.Collections.Generic.IEnumerable{NuGet.Common.PackagingLogMessage}")]
-[assembly: SuppressMessage("Build", "CA1822:Member GenerateWarningString does not access instance data and can be marked as static (Shared in VisualBasic)", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.Rules.NoRefOrLibFolderInPackageRule.GenerateWarningString(System.String[])~System.ValueTuple{System.String,System.String}")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'IEnumerable<PackagingLogMessage> PathTooLongRule.Validate(PackageArchiveReader builder)', validate parameter 'builder' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.Rules.PathTooLongRule.Validate(NuGet.Packaging.PackageArchiveReader)~System.Collections.Generic.IEnumerable{NuGet.Common.PackagingLogMessage}")]
 [assembly: SuppressMessage("Build", "CA1822:Member GetAllKeys does not access instance data and can be marked as static (Shared in VisualBasic)", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.Rules.ReferencesInNuspecMatchRefAssetsRule.GetAllKeys(System.Collections.Generic.IEnumerable{System.Linq.IGrouping{System.String,System.String}})~System.Collections.Generic.List{System.String}")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Accuracy Accuracy.Read(byte[] bytes)', validate parameter 'bytes' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.Packaging.Signing.Accuracy.Read(System.Byte[])~NuGet.Packaging.Signing.Accuracy")]
