+++ src/NuGet.Core/NuGet.Packaging/Rules/RuleSet.cs
@@ -25,7 +25,7 @@ namespace NuGet.Packaging.Rules
                 new PathTooLongRule(AnalysisResources.FilePathTooLongWarning),
                 new UnrecognizedLicenseIdentifierRule(AnalysisResources.UnrecognizedLicenseIdentifier),
                 new LicenseUrlDeprecationWarning(AnalysisResources.LicenseUrlDeprecationWarning),
-                new NoRefOrLibFolderInPackageRule(AnalysisResources.NoRefOrLibFolderInPackage),
+                new AccidentallyCompatibleWithAllFrameworksRule(),
                 new DependenciesGroupsForEachTFMRule(),
                 new UpholdBuildConventionRule(),
                 new ReferencesInNuspecMatchRefAssetsRule(),
